refactor(iai-callgrind): Move error into __internal mod
